// src/app/Models/client.model.ts
export interface StudentRegistration {
    firstName: string; // שם
    lastName: string;  // משפחה
    phone: string;     // טלפון
    idNumber: string;  // מספר זהות
    lesson: string;    // שיעור (lessonName)
    price: number;     // מחיר
    isPaid: boolean;   // האם שולם
}